requirements.txt
requests==2.32.3
python-dotenv==1.0.1
websocket-client==1.8.0
gspread==6.1.4
google-auth==2.37.0
telethon==1.37.0
